Allow scoping user/group to namespace
This allows scoping or restricting user/groups
to a namespace or namespaces. If namespace restrict
is turned on. It will block anyone without permission
to that particular namespace to do anthing other than
list.
