+++ config.json
@@ -1,4 +1,12 @@
 {
+	"RestrictNamespaceAccess": false,
+	"GroupNamespaces": {
+		"Engineers": ["all"],
+		"Admins": ["global", "kube-public"]
+	},
+	"UserNamespaces": {
+		"sulo.acharya": ["global"]
+	},
 	"BindDomain" : "example.com",
 	"BaseDN": "dc=example,dc=com",
 	"Filter": "(&(objectCategory=person)(objectClass=user)(samAccountName=%s))",
